requirements.txt
django==3.0.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-heroku==0.3.1
djoser==2.0.3
gunicorn==20.0.4
psycopg2-binary==2.8.4
python-dotenv==0.10.3
requests==2.22.0